refactor: wheelsize is now decimal (no additional cost and 27.5 is not 27 anymore)

+++ ams-desk-cs-backend/BikeFilters/Dtos/WheelSizeDto.cs
@@ -2,6 +2,6 @@
 
 public class WheelSizeDto
 {
-    public short Key { get; set; }
-    public short Value { get; set; }
+    public decimal Key { get; set; }
+    public decimal Value { get; set; }
 }
